% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describer.R
\name{describe}
\alias{describe}
\title{Create metadata for a data frame}
\usage{
describe(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a list
}
\description{
Prompts the user for short descriptions of a dataset as well as descriptions of the data contained in each of its columns
}
\examples{
\dontrun{
describe(example)
}
}
