% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describer.R
\name{docr}
\alias{docr}
\title{Create metadata string}
\usage{
docr(meta)
}
\arguments{
\item{meta}{metadata object created by \code{describe}}
}
\value{
a string that could be written to a .md document.
}
\description{
Create metadata string
}
\examples{
\dontrun{docr(out)}
}
