% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describer.R
\name{write_doc_csv}
\alias{write_doc_csv}
\title{Simultaneously save a data frame and it's metadata.}
\usage{
write_doc_csv(x, path, meta = "METADATA.md", desc = NULL, ...)
}
\arguments{
\item{x}{a data frame to write to disk}

\item{path}{path or connection to write to}

\item{meta}{name for the metadata document}

\item{desc}{if you've already saved a metadata object, you could pass this in here.  Otherwise, you'll be prompted to describe the data frame you are attempting to write.  Used mostly for testing purposes internally.}

\item{...}{other arguments passed to \code{\link[readr]{write_csv}}}
}
\value{

}
\description{
Simultaneously save a data frame and it's metadata.
}
\examples{
\dontrun{
library(here)
write_doc_csv(trees, here("data", "trees.csv"))
}
}
