% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describer.R
\name{write_with_meta}
\alias{write_with_meta}
\title{Simultaneously save a data frame and its metadata.}
\usage{
write_with_meta(
  x,
  path,
  meta = "METADATA.md",
  desc = NULL,
  write_func = c("csv", "csv2", "excel_csv", "excel_csv2", "tsv", "delim"),
  ...
)
}
\arguments{
\item{x}{a data frame to write to disk}

\item{path}{path or connection to write to}

\item{meta}{name for the metadata document}

\item{desc}{if you've already saved a metadata object, you could pass this in
here.  Otherwise, you'll be prompted to describe the data frame you are
attempting to write.  Used mostly for testing purposes internally.}

\item{write_func}{the suffix of the \code{\link[readr]{write_delim}} function
to use for writing the data frame.  Default is \code{"csv"} corresponding to
\code{\link[readr]{write_csv}}}

\item{...}{other arguments passed to \code{\link[readr]{write_csv}}}
}
\value{
returns the input \code{x} invisibly
}
\description{
Simultaneously save a data frame and its metadata.
}
\examples{
\dontrun{
library(here)
write_with_meta(trees, here("data", "trees.csv"))
}
}
