% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describer.R
\name{docr}
\alias{docr}
\title{Create metadata string}
\usage{
docr(x)
}
\arguments{
\item{x}{}
}
\value{
a string that could be written to a .md document.
}
\description{
Create metadata string
}
\examples{
docr(out)
}
